feat(custom/sidebar-width): Add selectors and storage key
